requirements.txt
fastapi==0.115.4
uvicorn==0.30.6

eth-account==0.13.7
mnemonic==0.21

google-cloud-firestore==2.21.0
